yt_companion_api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
Authlib==1.3.0
beautifulsoup4==4.12.3
blis==0.7.11
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.5
cymem==2.0.8
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.1
google-auth==2.29.0
grpcio==1.62.2
grpcio-health-checking==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
mutagen==1.47.0
nltk==3.8.1
numpy==1.26.4
packaging==24.0
preshed==3.0.9
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
pytube==15.0.0
regex==2024.4.28
requests==2.31.0
rsa==4.9
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
urllib3==2.2.1
validators==0.28.0
wasabi==1.1.2
weasel==0.3.4
weaviate-client==4.5.6
websockets==12.0
youtube-dl==2021.12.17
youtube-transcript-api==0.6.2
yt-dlp==2024.4.9
